Add custom font and style for recipes
Added a custom font to be used to recipe titles.
As of now, only the adobo recipe is styled. I've also added a .gitignore

I've used the same background-color for the index and the recipes.
Tried to use different fonts. Settled for now with Courier New,
monospace.

Soon, will also add the remaining recipes yet to be styled.